create scripts.servers module
