+++ scripts/build.py
@@ -30,6 +30,7 @@ import threading
 
 import python_utils
 from scripts import common
+from scripts import servers
 
 ASSETS_DEV_DIR = os.path.join('assets', '')
 ASSETS_OUT_DIR = os.path.join('build', 'assets', '')
@@ -669,7 +670,7 @@ def build_using_webpack(config_path):
     """
 
     python_utils.PRINT('Building webpack')
-    managed_webpack_compiler = common.managed_webpack_compiler(
+    managed_webpack_compiler = servers.managed_webpack_compiler(
         config_path=config_path, max_old_space_size=2400)
     with managed_webpack_compiler as p:
         p.wait()
